pyproject.toml
[tool.poetry]
name = "sangwon"
version = "0.1.0"
description = ""
authors = ["SW <aliwo@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
pydantic-settings = "^2.5.2"
pydantic = "^2.9.2"
tortoise-orm = {extras = ["asyncmy"], version = "^0.21.6"}
aerich = "^0.7.2"
cryptography = "^43.0.1"
httpx = "^0.27.2"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
mypy = "^1.11.2"
pytest = "^8.3.3"
coverage = "^7.6.1"


[tool.aerich]
tortoise_orm = "app.configs.database_settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
python_version = 3.12
strict = true

[tool.coverage.run]
source = ["app"]

[tool.coverage.report]
fail_under = 60
